% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPreview.R
\name{dataPreview}
\alias{dataPreview}
\title{This function will allow the user to create a time series plot of the data.}
\usage{
dataPreview(
  x,
  timepoints = c(2000, 2001, 2002, 2003, 2005),
  vertunits = "millions of variable",
  xAngle = 90
)
}
\arguments{
\item{x}{is the data which must be a rasterstack.}

\item{timepoints}{is a vector of numerical values representing your time points.}

\item{vertunits}{is a string that states the units of the y values.}

\item{xAngle}{is the orientation for the vertical axis labels. The default is 90.}
}
\value{
The output from \code{\link{dataPreview}}
}
\description{
This function will allow the user to create a time series plot of the data.
}
