% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presenceData.R
\name{presenceData}
\alias{presenceData}
\title{Create data for number of presence and change.}
\usage{
presenceData(x, nodata = -999)
}
\arguments{
\item{x}{is the data raster stack.}

\item{nodata}{is the interger that shows NA in the raster stack.}
}
\value{
The output from \code{\link{presenceData}}
}
\description{
Create data for number of presence and change.
}
