% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterstackData.R
\name{rasterstackData}
\alias{rasterstackData}
\title{Creates results that serve as input for the stackbarPlot function.}
\usage{
rasterstackData(
  x,
  timePoints = c(2000, 2001, 2002, 2003, 2005),
  unified = "yes",
  zeroabsence = "yes",
  categoryName = "variable",
  regionName = "region"
)
}
\arguments{
\item{x}{is the data, which must be a SpatRaste.}

\item{timePoints}{is a vector containing the time points under consideration.The default is  c(2000, 2001, 2002, 2003, 2005).}

\item{unified}{is a string, which can be "yes" or "no" only. If "yes," the change is a percentage of a region's
unified area; else, the change is a percentage of the  entire region under consideration.}

\item{zeroabsence}{is a string indicating whether 0 indicates basence or not}

\item{categoryName}{is a character representing the name of the category of interest.Default is "category"}

\item{regionName}{is a string or character the name of the study region.}
}
\value{
The output from \code{\link{rasterstackData}}
}
\description{
Creates results that serve as input for the stackbarPlot function.
}
