% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajPlot.R
\name{trajPlot}
\alias{trajPlot}
\title{Creates a map and pie chart of the trajectories of the category of interest.}
\usage{
trajPlot(
  input,
  axisShow = "no",
  categoryName = "forest",
  narrowPos = NA,
  scalePos = NA,
  scaleSize = 0.8,
  axisText = 1.2,
  axisLabel = 1.2,
  plotTitle = 1.3,
  legendTex = 1,
  xAxis = "Longitude (m)",
  yAxis = "Latitude (m)",
  downsample = TRUE
)
}
\arguments{
\item{input}{is the results from running the "raster_trajdata" function.}

\item{axisShow}{dictates whether the plot axes is shown or not. "no" means
no axes, whiles yes means the axes would be shown. The default is "no".}

\item{categoryName}{is a character representing the name of the category of interest.Default is "marsh".}

\item{narrowPos}{is a numeric value indicating the size of the north arrow.}

\item{scalePos}{where to align the scale bar. One of "bottomleft", "bottomright", "topleft", or "topright".}

\item{axisText}{is a numeric value controlling the size of the text on the horizontal and vertical ticks.}

\item{axisLabel}{is a numeric value controlling the size of the  horizontal and vertical labels}

\item{plotTitle}{is a numeric value controlling the size of the plot title.}

\item{legendTex}{is a numerical value controlling the size of the legend text.}

\item{xAxis}{is a character indicating label for the horizontal axis. default is "Longitude (m)".}

\item{yAxis}{is a character indicating label for the vertical axis. default is "Latitude (m)".}

\item{downsample}{is a string of "TRUE" or "FALSE" to down sample the spatial resolution of the plot.
"TRUE" down samples the plot.}

\item{dataEpsg}{is the projection of the current map. If extents are valid lat and lons,
the projection is assumed to be lat and lon (EPSG:4326), or Spherical Mercator otherwise.}

\item{scalesize}{is a numeric value indicating the size of the scale bar.}

\item{narrowSize}{is a numeric value indicating the size of the north arrow.}
}
\value{
The output from \code{\link{trajPlot}}
}
\description{
Creates a map and pie chart of the trajectories of the category of interest.
}
