% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterstackData.R
\name{rasterstackData}
\alias{rasterstackData}
\title{Creates results that serve as input for the stackbarPlot function.}
\usage{
rasterstackData(
  x,
  timePoints = c(2000, 2001, 2002, 2003, 2005),
  spatialextent = "unified",
  zeroabsence = "yes",
  annualchange = "yes",
  categoryName = "variable",
  regionName = "region"
)
}
\arguments{
\item{x}{is the data, which must be a SpatRaste. Default is "unified"}

\item{timePoints}{is a vector containing the time points under consideration.The default is  c(2000, 2001, 2002, 2003, 2005).}

\item{spatialextent}{is can be "unified", or any integer defined by the user. If "unified," the change is a percentage of unified region,
if user defined integer, the results are in the user’s desired units.
unified area; else, the change is a percentage of the  entire region under consideration.}

\item{zeroabsence}{is a string of "yes" or "no" indicating whether 0 indicates basence or not. default is "yes" meaning 0 means absence.}

\item{annualchange}{is a string of "yes" or "no". If "yes," results are expressed in annual change. Else, results are expressed in
the duration of the time interval. Defualt is "no"}

\item{categoryName}{is a character representing the name of the category of interest. Default is "category"}

\item{regionName}{is a string or character the name of the study region. Default is "region"}
}
\value{
The output from \code{\link{rasterstackData}}

The output from \code{\link{rasterstackData}}
}
\description{
Creates results that serve as input for the stackbarPlot function.
}
