% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rastertrajData.R
\name{rastertrajData}
\alias{rastertrajData}
\title{creates the data which serves as input for the "trajPlot" function.}
\usage{
rastertrajData(x, zeroabsence = "yes")
}
\arguments{
\item{x}{is the raster stack.}

\item{zeroabsence}{is a string of "yes" or "no" indicating if 0 means absence. # The unified size is the union of the locations where the category exists at any time point. The default is "yes".}
}
\value{
The output from \code{\link{rastertrajData}}
}
\description{
creates the data which serves as input for the "trajPlot" function.
}
